add miniconf package
